smart-light/requirements.txt
flask
paho-mqtt
PyJWT==2.8.0
